Add OpenSSL ports and streams
Add dependency on rust-openssl crate

Reorganize where the stream is wrapped in a BufStream

Add SSL ports to the server config. Idea for now is to use 0 for
invalid/unset.

Create enum to distinguish between SslStream and TcpStream. This enum
gets wrapped in a BufStream; it must therefore implement Read and
Write.

Create imap_ssl to wrap up a TcpStream in an SSL Stream if the
connection should perform the TLS handshake. Otherwise, wraps the
ordinary TcpStream in the server Stream enum. Add Option<SslAcceptor>
to the Server to perform this wrapping. If the Option is None, then
SSL is not configured for the server. This is the default value.